Changed naming schema for files with iRODS locations.
Renames the directory in line with teh directory name for the
iRODS restart files.

Made file names human readable where possible, ie followed
the convention adopted in this package. Added a suffix that contains
the name of the function class in case in future other iRODS loaders
start using this functionality.